Upgrade to electric v2.0.0-alpha
