+++ package.json
@@ -12,13 +12,7 @@
     "source-sans-pro": "^2.0.10"
   },
   "devDependencies": {
-    "electric": "^1.3.3",
-    "electric-quartz-components": "~0.0.1",
-    "gulp": "^3.9.0",
-    "gulp-connect": "^2.2.0",
-    "gulp-gh-pages": "^0.5.4",
-    "gulp-sass": "^2.0.1",
-    "metal-toggler": "^2.0.0",
-    "run-sequence": "^1.2.2"
+    "electric-quartz-components": "alpha",
+    "metal-toggler": "^2.0.0"
   }
 }
