--- src/layouts/blog.soy
@@ -1,53 +1,0 @@
-{namespace blog}
-
-/**
- *
- */
-{template .render}
-	{call main.render data="all"}
-		{param content kind="html"}
-			<div class="blog">
-				{call .mainPost data="all" /}
-				{call .olderPosts data="all" /}
-			</div>
-		{/param}
-	{/call}
-{/template}
-
-/**
- * @param content
- * @param page
- */
-{template .mainPost}
-	<article class="container">
-		<header>
-			<small>By {$page.author} <span>| {$page.date}</span></small>
-
-			<h3>{$page.title}</h3>
-		</header>
-
-		<div class="content">
-			{$content}
-		</div>
-	</article>
-{/template}
-
-/**
- * @param site
- */
-{template .olderPosts}
-	<div class="posts-list">
-		<div class="container">
-			<ul class="list-unstyled col-md-6 col-md-offset-3">
-				{foreach $page in $site.index.children[0].children}
-					<li class="text-center">
-						<a href="{$page.url}">
-							<small>By {$page.author} <span>| {$page.date}</span></small>
-							<p class="lead">{$page.title}</p>
-						</a>
-					</li>
-				{/foreach}
-			</ul>
-		</div>
-	</div>
-{/template}
