--- src/pages/blog/index.soy
@@ -1,26 +1,0 @@
----
-description: "Latest news."
-title: "Blog"
-type: "blog"
----
-
-{namespace pageBlogIndex}
-
-/**
- *
- */
-{template .render}
-{/template}
-
-/**
- * @param page
- */
-{template .soyweb}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta http-equiv="refresh" content="0; URL='{$page.redirect}'">
-</head>
-</html>
-{/template}
