+++ src/pages/docs/index.soy
@@ -6,14 +6,14 @@ description: "Documentation description"
 {namespace pageDocsIndex}
 
 /**
- * @param site
+ * @param page
  */
 {template .render}
 	{call main.render data="all"}
 		{param elementClasses: 'docs' /}
 		{param content kind="html"}
 			{call Sidebar.render}
-				{param section: $site.index.children[1] /}
+				{param section: $page /}
 			{/call}
 
 			{call .topics data="all" /}
@@ -22,6 +22,7 @@ description: "Documentation description"
 {/template}
 
 /**
+ * @param page
  * @param site
  */
 {template .topics}
@@ -65,7 +66,9 @@ description: "Documentation description"
 				<div class="row">
 					<div class="col-md-13 col-md-offset-3 col-xs-16">
 						<div class="row">
-							{foreach $topic in $site.index.children[1].children}
+							{foreach $childId in $page.childIds}
+								{let $topic: $page.children[$childId] /}
+
 								{if not $topic.hidden}
 									<div class="col-md-6 col-xs-16">
 										<a class="topic radial-out" href="{$topic.url}">
