+++ src/pages/docs/search.soy
@@ -14,7 +14,7 @@ hidden: true
 		{param elementClasses: 'docs' /}
 		{param content kind="html"}
 			{call Sidebar.render}
-				{param section: $site.index.children[1] /}
+				{param section: $site.index.children['docs'] /}
 			{/call}
 
 			<div class="sidebar-offset">
