+++ src/styles/main.scss
@@ -19,7 +19,7 @@
 
 // Font
 
-@import "font/source-sans-pro/source-sans-pro.css";
+@import "../vendor/source-sans-pro/source-sans-pro.css";
 
 body {
 	color: $brand-darker;
